backend/requirements.txt
Django==5.1.15
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
celery==5.4.0
django-celery-beat==2.7.0
redis==5.2.1
cloudinary==1.41.0
python-dotenv==1.0.1
django-cors-headers==4.6.0
django-ratelimit==4.1.0
requests==2.32.3
pytz==2024.2
stripe==10.8.0
setuptools==75.8.0
xhtml2pdf==0.2.11
WeasyPrint==62.3
Pillow==10.4.0

# AI Dependencies
openai==1.58.1
tiktoken==0.8.0
numpy==1.26.4
